requirements.txt
numpy>=1.18.2
tensorflow-gpu>=2.1.0
opencv-python>=4.1.2
scikit-image>=0.16.2
pandas>=1.0.3
matplotlib>=3.2.1
imageio>=2.4.1
